tmp: Rename module so overwritable app dir version can be imported
This is a temporary workaround to ko not supporting overwriting the app dir.
As go reasonably has no concept of soft forks, this hard forks the ko to
'unseenwizzard/ko' so the modified version allowing app dir overwrites
can be installed until a PR is created, reviewed and hopefully merged.